Move all test servers to a separate module, cleanup
